backend/reciPalApi/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
google-genai
yt_dlp
python-dotenv
ffmpeg
ffprobe